Предупреждай об удалении томов

+++ gandalf/containers/active/pi-hole/pi-hole/remove-volumes.sh
@@ -14,5 +14,12 @@ source ../../settings/settings-common.sh
 # Elevate privileges
 [ $UID -eq 0 ] || exec sudo --preserve-env=VERSION bash "$0" "$@"
 
+#https://github.com/alexanderfefelov/docker-backpack/blob/main/utils/cleanup/prune-all.sh
+read -p "WARNING: The data will be deleted. Press Y to continue: " -n 1 -r
+echo
+if [ "$REPLY" != "Y" ]; then
+  exit
+fi
+
 docker volume rm $CONTAINER_NAME-conf1
 docker volume rm $CONTAINER_NAME-conf2
